pyproject.toml
[project]
name = "scout"
version = "2.0.0"
description = "An enterprise context agent that navigates your company's knowledge graph"
requires-python = ">=3.12,<4"
readme = "README.md"
license-files = ["LICENSE"]
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
  "agno",
  "beautifulsoup4",
  "fastapi[standard]",
  "google-api-python-client",
  "google-auth",
  "httpx",
  "mcp",
  "openai",
  "openinference-instrumentation-agno",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "pgvector",
  "psycopg[binary]",
  "pypdf",
  "python-docx",
  "pyyaml",
  "sqlalchemy",
]

[project.optional-dependencies]
dev = ["mypy", "ruff", "types-PyYAML"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scout*", "scout.sources*", "scout.compile*", "db*", "app*", "evals*", "context*"]

[tool.ruff]
line-length = 120
exclude = [".venv*"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = [
  "pgvector.*",
  "setuptools.*",
  "agno.*",
  "httpx.*",
  "mcp.*",
  "parallel.*",
  "google.*",
  "googleapiclient.*",
  "slack_sdk.*",
  "yaml.*",
  "pypdf.*",
  "docx.*",
  "bs4.*",
]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true
